Module Private.Cmd

Executing commands.

type t = string list

The type for command line fragments

val find : string list -> string option

find bins looks for the first binary in bins in PATH.

val run : t -> (unit, int * string) result

run cmd runs the command cmd. In case of error returns the command's exit code and an error message.

val read : t -> (string, int * string) result

run cmd runs the command cmd and reads its standard input. In case of error returns the command's exit code and an error message.