Module Private.Top

Abstracts away the OCaml Toploop API.

Toplevel directives

val add_inc : string -> (unit, string) result

add_inc dir adds dir to the include path.

val rem_inc : string -> (unit, string) result

rem_inc dir removes dir from the include path.

val load_ml : string -> (unit, string) result

load_ml ml loads the source file ml.

val load_obj : string -> (unit, string) result

load_obj obj loads the object file obj.

val is_nat : unit -> bool

is_nat () is true if the current toplevel implementation is ocamlnat.

Topdirs implementations

module type TOPDIRS = sig ... end

The type for Topdir module implementations.

val set_topdirs : is_nat:bool -> (module TOPDIRS) -> unit

set_topdirs ~is_nat t sets the topdirs implementation to t and indicates with is_nat if this ocamlnat's implementation.