Module Omod_ocamlc.Cma

cma files

type t

The type for cma files.

val read : Omod.fpath -> (t, string) result

read_cma f reads a cma ile from f.

val name : t -> string

name cma is the archive name of cma.

val cmos : t -> Cmo.t list

cmos cma are the cmos contained in the archive.

val custom : t -> bool

custom cma is true if it requires custom mode linking.

val custom_cobjs : t -> string list

custom_cobjs are C objects files needed for custom mode linking.

val custom_copts : t -> string list

custom_copts are extra options passed for custom mode linking.

val dllibs : t -> string list

dllibs are dynamically loaded C libraries for ocamlrun dynamic linking.