Module Omod_ocamlc.Cmxa

cmxa files

type t

The type for cmxa files

val read : Omod.fpath -> (t, string) result

read_cmxa f reads a cmxa file from f.

val name : t -> string

name cmxa is the archive name of cmxa.

val cmxs : t -> Cmx.t list

cmxs cmxa are the cmxs contained in the archive.

val cobjs : t -> string list

cobjs cmxa are C objects files needed for linking.

val copts : t -> string list

copts cmxa are options for the C linker.