Module Omod_support.Codec

Value serializers.

The underlying implementation uses Marshal with magic numbers that depend on omod and OCaml's versions.

type 'a t

The type for codec of values of type 'a.

val v : id:string -> 'a t

v ~id is a new codec with identifier id (part of the magic number).

val write : 'a t -> Omod.fpath -> 'a -> (unit, string) result

write c file v writes value v to file using codec c.

val read : 'a t -> Omod.fpath -> ('a, string) result

read c file reads a value from file using codec c.