Module Omod_support.Conf

Omod configuration.

type t

The type for configuration.

val v : ?cache:Omod.fpath -> ?libdir:Omod.fpath -> unit -> (t, string) result

v ~libdir ~cache () is a configuration with given cache and libdir. If unspecified they are discovered.

val cache : t -> Omod.fpath

cache c is c's cache directory.

val libdir : t -> Omod.fpath

libdir c is c's library directory.

val pp : Format.formatter -> t -> unit

pp formats configurations.

Environment variables

val cache_env : string

cache_env is the environment variable that can be used to define an cache directory.

val libdir_env : string

libdir_env is the environment variable that can be used to define a libdir.