Module Omod_support.Log

Logging.

type t = {

f : a. ('a, Format.formatter, unit) Pervasives.format -> 'a;

}

The type for logging functions.

val nil : t

nil is a logging function that drops logging.

val std : t

std is a logging function that format on Format.std_formatter and prepends messages by the executable basename.

val err : t

err is like std but formats on Format.err_formatter.

val time : t -> string -> (unit -> 'a) -> 'a

time l label f logs the processor time of f () on l.